casted targets to long

+++ RNN.py
@@ -106,7 +106,7 @@ class RNN(nn.Module):
                 # print(outputs[1:, :, :].shape)
                 # print(targets.transpose(0,2).transpose(1,2).shape)
                 outputs = outputs[1:, :, :]
-                targets = targets.transpose(0, 2).transpose(1, 2)
+                targets = targets.transpose(0, 2).transpose(1, 2).long()
                 # print(targets)
                 for bat in range(batch_size):
                     loss += loss_function(outputs[1:, bat, :], targets[:, bat, :].squeeze(1))
